Extract slider code to a separate component
